move likelihoods to dedicated module
