+++ darts/utils/likelihood/likelihood_models.py
@@ -57,7 +57,7 @@ from torch.distributions import Weibull as _Weibull
 from torch.distributions.kl import kl_divergence
 
 from darts.logging import raise_if_not
-from darts.utils.likelihood_utils import BaseLikelihood, LikelihoodType
+from darts.utils.likelihood.likelihood import BaseLikelihood, LikelihoodType
 
 # TODO: Table on README listing distribution, possible priors and wiki article
 from darts.utils.utils import (
