+++ docs/userguide/hyperparameter_optimization.md
@@ -27,7 +27,7 @@ from darts.dataprocessing.transformers import Scaler
 from darts.datasets import AirPassengersDataset
 from darts.metrics import smape
 from darts.models import TCNModel
-from darts.utils.likelihood_models import GaussianLikelihood
+from darts.utils.likelihood.likelihood_models import GaussianLikelihood
 
 # load data
 series = AirPassengersDataset().load().astype(np.float32)
@@ -41,16 +41,18 @@ scaler = Scaler(MaxAbsScaler())
 train = scaler.fit_transform(train)
 val = scaler.transform(val)
 
+
 #  workaround found in https://github.com/Lightning-AI/pytorch-lightning/issues/17485
 # to avoid import of both lightning and pytorch_lightning
 class PatchedPruningCallback(optuna.integration.PyTorchLightningPruningCallback, Callback):
     pass
 
+
 # define objective function
 def objective(trial):
     # select input and output chunk lengths
     in_len = trial.suggest_int("in_len", 12, 36)
-    out_len = trial.suggest_int("out_len", 1, in_len-1)
+    out_len = trial.suggest_int("out_len", 1, in_len - 1)
 
     # Other hyperparameters
     kernel_size = trial.suggest_int("kernel_size", 2, 5)
@@ -108,10 +110,9 @@ def objective(trial):
         save_checkpoints=True,
     )
 
-
     # when validating during training, we can use a slightly longer validation
     # set which also contains the first input_chunk_length time steps
-    model_val_set = scaler.transform(series[-(VAL_LEN + in_len) :])
+    model_val_set = scaler.transform(series[-(VAL_LEN + in_len):])
 
     # train the model
     model.fit(
